+++ packages/x-charts/src/ScatterChart/index.ts
@@ -5,3 +5,5 @@ export * from './ScatterPlot';
 export * from './Scatter';
 
 export * from './ScatterMarker';
+
+export * from './Scatter.types';
